Update Java examples
Build out the start of new set of Java Examples that are easier to work
with. Make it a single Gradle application so it's easy to build. Unify
the command line arguments so one example is as easy as another to
execute.

Add more README files to help explain overall as well as sections of
examples for clarity.